gcp/split-tiff-function/requirements.txt
google-cloud-secret-manager
google-cloud-storage
rasterio
Shapely

